Create ConfigHelper class
